Add FileStorageService interface and implementation
